requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
filelock==3.20.0
fsspec==2025.12.0
gitdb==4.0.12
gitpython==3.1.45
idna==3.11
imagehash==4.3.1
jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.13.0
networkx==3.6.1
numpy==2.3.5
opencv-python==4.10.0.84
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.2
pyarrow==22.0.0
pydeck==0.9.1
pygments==2.19.2
pymysql==1.1.0
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2025.2
pywavelets==1.9.0
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sqlalchemy==2.0.30
streamlit==1.52.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tornado==6.5.3
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2

